not ready. check in for backup

+++ routes/healthRecord.js
@@ -1,20 +1,22 @@
 const router = require('express').Router();
 const { isAuthorized, isPremiumUser } = require('../middleware/auth');
-const { createHealthRecord, getHealthRecordByUserId, updateHealthRecord, deleteRecordById, deleteAllRecordsByUserId } = require('../daos/healthRecord');
+const { createHealthRecord, getHealthRecordByUserId, updateHealthRecord, deleteRecordById, deleteAllRecordsByUserId, getHealthRecordAll } = require('../daos/healthRecord');
 
 router.post("/create", isAuthorized, async (req, res, next) => {
-    try {
+    try {        
+        //console.log("querystring in Post", req.query)
 
+        if (req.query.errorMode == 'y')
+            throw new Error("error test")
+        
         const healthRecordObj = {
             userId: req.userData._id,
-            sleepHours: req.body.sleepHour,
+            sleepHours: req.body.sleepHours,
             weight: req.body.weight,
             highBP: req.body.highBP,
             lowBP: req.body.lowBP,
-            dateTime: req.body.dateTime
-        }
-
-        //console.log('healthRecordObj' , healthRecordObj);
+            dateTime: new Date(req.body.dateTime)
+        }        
         const created = await createHealthRecord(healthRecordObj);
         return res.status(200).json(created);
     } catch (e) {
@@ -24,6 +26,9 @@ router.post("/create", isAuthorized, async (req, res, next) => {
 
 router.put("/:id", isAuthorized, async (req, res, next) => {
     try {
+        //console.log("querystring in Put", req.query)
+        if (req.query.errorMode == 'y')
+            throw new Error("error test")
 
         const healthRecordObj = {
             recordId: req.params.id,
@@ -34,27 +39,21 @@ router.put("/:id", isAuthorized, async (req, res, next) => {
             lowBP: req.body.lowBP,
             dateTime: req.body.dateTime
         }
-
         const updated = await updateHealthRecord(healthRecordObj)
         return res.status(200).json(updated)
-    } catch (e) {
+    } catch (e) {        
+        console.log('fall in catch routine', e.message);
         next(e)
     }
 });
 
 router.get("/", isAuthorized, async (req, res, next) => {
     try {
+        if (req.query.errorMode == 'y')
+            throw new Error("error test")
+
         const records = await getHealthRecordByUserId(req.userData._id);
-
-        //console.log('req.userData._id', req.userData._id);
-        // if (req.userData.roles.includes('admin') || req.userData._id === order.userId.toString()) 
-        //     return res.json(order);
-
-        if (records) {
-        //    console.log('records', JSON.stringify(records))
-            return res.json(records)
-        }
-        res.sendStatus(404);
+        return res.json(records);        
     } catch (e) {
         next(e);
     }
@@ -62,6 +61,9 @@ router.get("/", isAuthorized, async (req, res, next) => {
 
 router.delete("/:id", isAuthorized, async (req, res, next) => {
     try {
+        if (req.query.errorMode == 'y')
+            throw new Error("error test")
+
         const item = await deleteRecordById(req.params.id);
         return res.json(item);
     } catch (e) {
@@ -69,9 +71,11 @@ router.delete("/:id", isAuthorized, async (req, res, next) => {
     }
 });
 
-
 router.delete("/", isAuthorized, async (req, res, next) => {
     try {
+        if (req.query.errorMode == 'y')
+            throw new Error("error test")
+
         const records = await deleteAllRecordsByUserId(req.userData._id);
         return res.json(records);
     } catch (e) {
